Create paging module and move paging classes to module
