+++ core/paging/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+  id("plugin.tvmaniac.multiplatform")
+}
+
+kotlin {
+  sourceSets {
+    commonMain.dependencies {
+      implementation(libs.coroutines.core)
+      implementation(libs.paging.common)
+    }
+  }
+}
